fix: subdivide DTO Level
