--- src/main/java/com/game/rmt/domain/game/dto/GameListResponse.java
@@ -1,14 +1,0 @@
-package com.game.rmt.domain.game.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.util.List;
-
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor
-public class GameListResponse {
-    private List<GameDTO> gameList;
-}
