+++ src/main/java/com/game/rmt/domain/game/dto/request/GameSearchFilter.java
@@ -1,4 +1,4 @@
-package com.game.rmt.domain.game.dto;
+package com.game.rmt.domain.game.dto.request;
 
 import com.game.rmt.global.parent.Filter;
 import lombok.AllArgsConstructor;
