+++ src/main/java/com/game/rmt/domain/game/dto/response/GameListResponse.java
@@ -1,5 +1,6 @@
-package com.game.rmt.domain.platform.dto;
+package com.game.rmt.domain.game.dto.response;
 
+import com.game.rmt.domain.game.dto.GameDTO;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -9,6 +10,6 @@ import java.util.List;
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor
-public class PlatformListResponse {
-    private List<PlatformDTO> platformList;
+public class GameListResponse {
+    private List<GameDTO> gameList;
 }
