+++ src/main/java/com/game/rmt/domain/game/service/GameService.java
@@ -2,8 +2,8 @@ package com.game.rmt.domain.game.service;
 
 import com.game.rmt.domain.game.domain.Game;
 import com.game.rmt.domain.game.dto.GameDTO;
-import com.game.rmt.domain.game.dto.GameSearchFilter;
-import com.game.rmt.domain.game.dto.NewGameRequest;
+import com.game.rmt.domain.game.dto.request.GameSearchFilter;
+import com.game.rmt.domain.game.dto.request.NewGameRequest;
 import com.game.rmt.domain.game.repository.GameRepository;
 import com.game.rmt.domain.platform.domain.Platform;
 import com.game.rmt.domain.platform.service.PlatformService;
