+++ src/main/java/com/game/rmt/domain/product/service/ProductService.java
@@ -4,9 +4,9 @@ import com.game.rmt.domain.game.domain.Game;
 import com.game.rmt.domain.game.service.GameService;
 import com.game.rmt.domain.platform.service.PlatformService;
 import com.game.rmt.domain.product.domain.Product;
-import com.game.rmt.domain.product.dto.NewProductRequest;
+import com.game.rmt.domain.product.dto.request.NewProductRequest;
 import com.game.rmt.domain.product.dto.ProductDTO;
-import com.game.rmt.domain.product.dto.ProductSearchFilter;
+import com.game.rmt.domain.product.dto.request.ProductSearchFilter;
 import com.game.rmt.domain.product.repository.ProductRepository;
 import com.game.rmt.global.errorhandler.exception.BadRequestException;
 import com.game.rmt.global.errorhandler.exception.NotFoundException;
