+++ src/main/java/com/game/rmt/domain/statistics/controller/StatisticsController.java
@@ -1,8 +1,8 @@
 package com.game.rmt.domain.statistics.controller;
 
-import com.game.rmt.domain.statistics.dto.MonthlyEachGameResponse;
-import com.game.rmt.domain.statistics.dto.MonthlyGameRequest;
-import com.game.rmt.domain.statistics.dto.MonthlyPlatformRequest;
+import com.game.rmt.domain.statistics.dto.response.MonthlyEachGameResponse;
+import com.game.rmt.domain.statistics.dto.request.MonthlyGameRequest;
+import com.game.rmt.domain.statistics.dto.request.MonthlyPlatformRequest;
 import com.game.rmt.domain.statistics.service.StatisticsService;
 import lombok.AllArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
