+++ src/main/java/com/game/rmt/domain/statistics/dto/response/MonthlyEachGameResponse.java
@@ -1,5 +1,6 @@
-package com.game.rmt.domain.statistics.dto;
+package com.game.rmt.domain.statistics.dto.response;
 
+import com.game.rmt.domain.statistics.dto.MonthlyStaticsDTO;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
