+++ src/main/java/com/game/rmt/domain/statistics/service/StatisticsService.java
@@ -4,9 +4,9 @@ import com.game.rmt.domain.game.domain.Game;
 import com.game.rmt.domain.game.service.GameService;
 import com.game.rmt.domain.platform.domain.Platform;
 import com.game.rmt.domain.platform.service.PlatformService;
-import com.game.rmt.domain.statistics.dto.MonthlyEachGameResponse;
-import com.game.rmt.domain.statistics.dto.MonthlyGameRequest;
-import com.game.rmt.domain.statistics.dto.MonthlyPlatformRequest;
+import com.game.rmt.domain.statistics.dto.response.MonthlyEachGameResponse;
+import com.game.rmt.domain.statistics.dto.request.MonthlyGameRequest;
+import com.game.rmt.domain.statistics.dto.request.MonthlyPlatformRequest;
 import com.game.rmt.domain.statistics.dto.MonthlyStaticsDTO;
 import com.game.rmt.domain.statistics.repository.custom.StaticsRepository;
 import lombok.RequiredArgsConstructor;
