+++ src/test/java/com/game/rmt/domain/product/service/ProductServiceTest.java
@@ -7,9 +7,9 @@ import com.game.rmt.domain.platform.domain.Platform;
 import com.game.rmt.domain.platform.repository.PlatformRepository;
 import com.game.rmt.domain.platform.service.PlatformService;
 import com.game.rmt.domain.product.domain.Product;
-import com.game.rmt.domain.product.dto.NewProductRequest;
+import com.game.rmt.domain.product.dto.request.NewProductRequest;
 import com.game.rmt.domain.product.dto.ProductDTO;
-import com.game.rmt.domain.product.dto.ProductSearchFilter;
+import com.game.rmt.domain.product.dto.request.ProductSearchFilter;
 import com.game.rmt.domain.product.repository.ProductRepository;
 import com.game.rmt.global.errorhandler.exception.BadRequestException;
 import com.game.rmt.global.errorhandler.exception.ErrorCode;
